fix(android) upgrade to v1.3.0 and change float to Double

+++ android/src/main/kotlin/video/api/flutter/player/FlutterPlayerInterface.kt
@@ -24,8 +24,8 @@ interface FlutterPlayerInterface {
     fun getVolume(textureId: Long): Float
     fun setVolume(textureId: Long, volume: Float)
     fun getVideoSize(textureId: Long): Size?
-    fun getPlaybackSpeed(textureId: Long): Float
-    fun setPlaybackSpeed(textureId: Long, playbackSpeed: Float)
+    fun getPlaybackSpeed(textureId: Long): Double
+    fun setPlaybackSpeed(textureId: Long, playbackSpeed: Double)
 
     fun play(textureId: Long)
     fun pause(textureId: Long)
